backend/requirements.txt
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
backcall==0.2.0
bce-python-sdk==0.9.64
beautifulsoup4==4.14.3
bleach==4.1.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
chardet==7.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
cryptography==46.0.3
cssselect2==0.8.0
decorator==5.2.1
defusedxml==0.7.1
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
django-jazzmin==3.0.1
django-unfold==0.73.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
docopt==0.6.2
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.25.2
fonttools==4.61.1
fsspec==2026.2.0
future==1.0.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
imagesize==2.0.0
imutils==0.5.4
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
modelscope==1.35.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.6.1
numpy==2.2.6
oauthlib==3.3.1
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.4.0
paddlepaddle==3.3.0
paddlex==3.4.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pdf2image==1.17.0
pickleshare==0.7.5
pillow==12.0.0
pipreqs==0.5.0
platformdirs==4.5.1
prettytable==3.17.0
prompt_toolkit==3.0.52
protobuf==7.34.0
psutil==7.2.2
psycopg2-binary==2.9.11
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.12.1
Pygments==2.19.2
PyJWT==2.10.1
pypdfium2==5.6.0
pyphen==0.17.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python3-openid==3.2.0
pytils==0.4.4
pytz==2025.2
PyYAML==6.0.2
pyzbar==0.1.9
pyzmq==27.1.0
qrcode==8.2
referencing==0.37.0
reportlab==4.4.7
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
ruamel.yaml==0.19.1
safetensors==0.7.0
setuptools==82.0.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
soupsieve==2.8.1
sqlparse==0.5.4
stack-data==0.6.3
tinycss2==1.5.1
tinyhtml5==2.0.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.12.0
urllib3==2.6.2
wcwidth==0.2.14
weasyprint==68.1
webencodings==0.5.1
yarg==0.1.9
zopfli==0.4.0
pyTelegramBotAPI==4.23.0
